requirements.txt
pandas>=2.1.4,<2.2.0
numpy>=1.26.3,<1.27.0
requests==2.31.0
python-dotenv==1.0.0
pyinstaller==6.3.0